+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This document details the changes between each release.
 
 * OSC bundle support, via a simple `OSCBundle` container class.
 * A `LiteOSCParser::clear()` function that empties the message.
+* A test for out-of-range matching.
 
 ### Changed
 
@@ -18,6 +19,8 @@ This document details the changes between each release.
 ### Fixed
 
 * The constructors now check if the initial memory allocation failed.
+* The `match` function now returns the correct value if the offset is equal
+  to the address length and the pattern to be matched is empty.
 
 ## LiteOSCParser v1.0.0